+++ src/defines.ts
@@ -1,12 +1,3 @@
-// LN_SCALE_MIN..LN_SCALE_MAX define the internal scale range of for Gsplats,
-// covering approx 0.0001..8000 in range with discrete steps 7% apart.
-// The value "0" is reserved for truly flat scales, indicating a 2DGS.
-// If these values are changed, the corresponding values in splatDefines.glsl
-// must also be updated to match.
-
-export const LN_SCALE_MIN = -12.0;
-export const LN_SCALE_MAX = 9.0;
-
 // Gsplats are stored in textures that are 2^11 x 2^11 x up to 2^11
 // Most WebGL2 implementations support 2D textures up to 2^12 x 2^12 (max 16M Gsplats)
 // 2D array textures and 3D textures up to 2^11 x 2^11 x 2^11 (max 8G Gsplats),